xmhf: xmhf-core/xcprimeon: test to ensure x86-64 ABI compliance; invoke xmhf_xcphandler_arch_hub to ensure 1st parameter goes into rdi and the second into rsi as per the ABI;success.

+++ xmhf-slabs/xmhf-core/xcexhub/arch/x86-vmx-x86pc/xcexhub-x86vmxx86pc-code.c
@@ -100,7 +100,7 @@ __attribute__(( section(".slab_trampoline") )) static void xmhf_xcphandler_arch_
 //==========================================================================================
 
 //exception handler hub
-__attribute__(( section(".slab_trampoline") )) void xmhf_xcphandler_arch_hub(u32 vector, void *exdata){
+__attribute__(( section(".slab_trampoline") )) void xmhf_xcphandler_arch_hub(u64 vector, void *exdata){
     struct regs *r = (struct regs *)exdata;
 
 	switch(vector){
